Modify: change WhiteContainer to tsx
