--- src/components/common/WhiteContainer.jsx
@@ -1,47 +1,0 @@
-import React from "react";
-import RLayout from "components/common/RLayout";
-import PropTypes from "prop-types";
-
-const WhiteContainer = (props) => {
-  const box = (
-    <div
-      style={{
-        margin: props.margin,
-        boxShadow:
-          "0px 1.5px 1px -0.5px rgba(110, 54, 120, 0.05), 0px 2.5px 1px -0.5px rgba(110, 54, 120, 0.03), 0px 2px 3px -1px rgba(110, 54, 120, 0.11)",
-        borderRadius: "12px",
-        overflow: "hidden",
-      }}
-    >
-      <div
-        style={{
-          position: "relative",
-          padding: props.padding,
-          background: "white",
-        }}
-      >
-        {props.children}
-      </div>
-    </div>
-  );
-
-  if (props.marginAuto) {
-    return <RLayout.R1>{box}</RLayout.R1>;
-  }
-  return box;
-};
-
-WhiteContainer.propTypes = {
-  // FIXME specify type
-  children: PropTypes.any,
-  padding: PropTypes.any,
-  marginAuto: PropTypes.bool,
-  margin: PropTypes.string,
-};
-WhiteContainer.defaultProps = {
-  padding: "24px",
-  margin: "0 0 15px",
-  marginAuto: false,
-};
-
-export default WhiteContainer;
